+++ pkg/util/k8sutil/util.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -50,8 +50,12 @@ const (
 	LabelKeyArangoActive = "deployment.arangodb.com/active"
 	// LabelValueArangoActive is the value of the label used to mark members as active.
 	LabelValueArangoActive = "true"
-	// AppName is the fixed value for the "app" label
+	// LabelKeyArangoMLStatefulSet is the key of the label used to define k8s StatefulSet for ML Extension
+	LabelKeyArangoMLStatefulSet = "ml.arangodb.com/statefulset"
+	// AppName is the value for the "app" label
 	AppName = "arangodb"
+	// AppArangoML is the value for the "app" label
+	AppArangoML = "arangoml"
 )
 
 // AddOwnerRefToObject adds given owner reference to given object